+++ CHANGES.md
@@ -3,7 +3,7 @@
 
 ## 0.3.2 / 2016-March-??
 
-* [TODO] Use function `cambium.core/stringify-key` to convert MDC key to string (can be redefined)
+* Use function `cambium.core/stringify-key` to convert MDC key to string (can be redefined)
 
 
 ## 0.3.1 / 2016-March-10